+++ rs/tests/Cargo.toml
@@ -54,6 +54,7 @@ ic-ckbtc-minter = { path = "../bitcoin/ckbtc/minter" }
 ic-config = { path = "../config" }
 ic-constants = { path = "../constants" }
 ic_consensus_system_test_utils = { path = "./consensus/utils" }
+ic_consensus_threshold_sig_system_test_utils = { path = "./consensus/tecdsa/utils" }
 ic-crypto-sha2 = { path = "../crypto/sha2" }
 ic-crypto-test-utils-reproducible-rng = { path = "../crypto/test_utils/reproducible_rng" }
 ic-crypto-tree-hash = { path = "../crypto/tree_hash" }
@@ -199,20 +200,8 @@ name = "ic-systest-rotate-ecdsa-idkg-key"
 path = "consensus/orchestrator/rotate_ecdsa_idkg_key_test.rs"
 
 [[bin]]
-name = "ic-systest-tecdsa-add-nodes"
-path = "consensus/tecdsa/tecdsa_add_nodes_test.rs"
-
-[[bin]]
-name = "ic-systest-tschnorr-message-sizes"
-path = "consensus/tecdsa/tschnorr_message_sizes_test.rs"
-
-[[bin]]
 name = "ic-systest-tecdsa-key-rotation"
 path = "consensus/tecdsa/tecdsa_key_rotation_test.rs"
-
-[[bin]]
-name = "ic-systest-tecdsa-remove-nodes"
-path = "consensus/tecdsa/tecdsa_remove_nodes_test.rs"
 
 [[bin]]
 name = "ic-systest-tecdsa-signature-life-cycle"
@@ -357,10 +346,6 @@ path = "consensus/safety_test.rs"
 [[bin]]
 name = "ic-systest-system-api-security-test"
 path = "execution/system_api_security_test.rs"
-
-[[bin]]
-name = "ic-systest-tecdsa-complaint-test"
-path = "consensus/tecdsa/tecdsa_complaint_test.rs"
 
 [[bin]]
 name = "ic-systest-max-ingress-payload-size-test"
