+++ rs/tests/common.bzl
@@ -95,6 +95,7 @@ DEPENDENCIES = [
     "//rs/test_utilities/time",
     "//rs/test_utilities/types",
     "//rs/tests/consensus/utils",
+    "//rs/tests/consensus/tecdsa/utils",
     "//rs/tests/driver:ic-system-test-driver",
     "//rs/tests/test_canisters/message:lib",
     "//rs/tree_deserializer",
