+++ rs/tests/consensus/tecdsa/Cargo.toml
@@ -8,16 +8,25 @@ documentation.workspace = true
 
 [dependencies]
 anyhow = { workspace = true }
+canister-test = { path = "../../../rust_canisters/canister_test" }
 chrono = { workspace = true }
 futures = { workspace = true }
 ic_consensus_system_test_utils = { path = "../utils" }
-ic-management-canister-types = { path = "../../..//types/management_canister_types" }
+ic_consensus_threshold_sig_system_test_utils = { path = "./utils" }
+ic-config = { path = "../../../config" }
+ic-constants = { path = "../../../constants" }
+ic-base-types = { path = "../../../types/base_types" }
+ic-management-canister-types = { path = "../../../types/management_canister_types" }
+ic-nns-constants = { path = "../../../nns/constants" }
+ic-nns-governance = { path = "../../../nns/governance" }
 ic-registry-subnet-features = { path = "../../../registry/subnet_features" }
 ic-registry-subnet-type = { path = "../../../registry/subnet_type" }
 ic-sys = { path = "../../../sys" }
 ic-system-test-driver = { path = "../../driver" }
 ic-types = { path = "../../../types/types" }
 reqwest = { workspace = true }
+rand = { workspace = true }
+rand_chacha = { workspace = true }
 serde_json = { workspace = true }
 slog = { workspace = true }
 tests = { path = "../.." }
@@ -28,3 +37,50 @@ url = { workspace = true }
 name = "ic-systest-tecdsa-performance-test"
 path = "tecdsa_performance_test.rs"
 
+[[bin]]
+name = "ic-systest-tecdsa-add-nodes"
+path = "consensus/tecdsa/tecdsa_add_nodes_test.rs"
+
+[[bin]]
+name = "ic-systest-tschnorr-message-sizes"
+path = "consensus/tecdsa/tschnorr_message_sizes_test.rs"
+
+# [[bin]]
+# name = "ic-systest-tecdsa-key-rotation"
+# path = "consensus/tecdsa/tecdsa_key_rotation_test.rs"
+
+[[bin]]
+name = "ic-systest-tecdsa-remove-nodes"
+path = "consensus/tecdsa/tecdsa_remove_nodes_test.rs"
+
+# [[bin]]
+# name = "ic-systest-tecdsa-signature-life-cycle"
+# path = "consensus/tecdsa/tecdsa_signature_life_cycle_test.rs"
+
+# [[bin]]
+# name = "ic-systest-tecdsa-signature-fails-without-cycles"
+# path = "consensus/tecdsa/tecdsa_signature_fails_without_cycles_test.rs"
+
+# [[bin]]
+# name = "ic-systest-tecdsa-signature-from-nns-without-cycles"
+# path = "consensus/tecdsa/tecdsa_signature_from_nns_without_cycles_test.rs"
+
+# [[bin]]
+# name = "ic-systest-tecdsa-signature-from-other-subnet"
+# path = "consensus/tecdsa/tecdsa_signature_from_other_subnet_test.rs"
+
+# [[bin]]
+# name = "ic-systest-tecdsa-signature-same-subnet"
+# path = "consensus/tecdsa/tecdsa_signature_same_subnet_test.rs"
+
+# [[bin]]
+# name = "ic-systest-tecdsa-signature-timeout"
+# path = "consensus/tecdsa/tecdsa_signature_timeout_test.rs"
+
+# [[bin]]
+# name = "ic-systest-tecdsa-two-signing-subnets"
+# path = "consensus/tecdsa/tecdsa_two_signing_subnets_test.rs"
+
+[[bin]]
+name = "ic-systest-tecdsa-complaint-test"
+path = "consensus/tecdsa/tecdsa_complaint_test.rs"
