+++ rs/tests/consensus/tecdsa/utils/BUILD.bazel
@@ -1,0 +1,43 @@
+load("@rules_rust//rust:defs.bzl", "rust_library")
+
+package(default_visibility = ["//rs:system-tests-pkg"])
+
+rust_library(
+    name = "utils",
+    testonly = True,
+    srcs = glob(["src/**/*.rs"]),
+    crate_name = "ic_consensus_threshold_sig_system_test_utils",
+    deps = [
+        # Keep sorted.
+        "//rs/canister_client",
+        "//rs/config",
+        "//rs/constants",
+        "//rs/nervous_system/common/test_keys",
+        "//rs/nns/cmc",
+        "//rs/nns/common",
+        "//rs/nns/constants",
+        "//rs/nns/governance",
+        "//rs/nns/test_utils",
+        "//rs/prep",
+        "//rs/registry/canister",
+        "//rs/registry/subnet_features",
+        "//rs/registry/subnet_type",
+        "//rs/rust_canisters/canister_test",
+        "//rs/tests/driver:ic-system-test-driver",
+        "//rs/tests/test_canisters/message:lib",
+        "//rs/types/base_types",
+        "//rs/types/management_canister_types",
+        "//rs/types/types",
+        "//rs/types/types_test_utils",
+        "@crate_index//:anyhow",
+        "@crate_index//:candid",
+        "@crate_index//:ed25519-dalek",
+        "@crate_index//:ic-agent",
+        "@crate_index//:k256",
+        "@crate_index//:schnorr_fun",
+        "@crate_index//:serde",
+        "@crate_index//:sha2",
+        "@crate_index//:slog",
+        "@crate_index//:tokio",
+    ],
+)
