+++ rs/tests/src/orchestrator/utils/subnet_recovery.rs
@@ -2,11 +2,6 @@ use std::collections::BTreeMap;
 use std::time::Duration;
 
 use crate::orchestrator::utils::upgrade::assert_assigned_replica_version;
-use crate::tecdsa::{
-    add_chain_keys_with_timeout_and_rotation_period, create_new_subnet_with_keys,
-    empty_subnet_update, execute_update_subnet_proposal, get_public_key_with_retries,
-    get_signature_with_logger, verify_signature,
-};
 use anyhow::bail;
 use candid::Principal;
 use canister_test::Canister;
@@ -17,6 +12,11 @@ use ic_consensus_system_test_utils::{
         can_read_msg, can_store_msg, cannot_store_msg, cert_state_makes_progress_with_retries,
     },
     ssh_access::execute_bash_command,
+};
+use ic_consensus_threshold_sig_system_test_utils::{
+    add_chain_keys_with_timeout_and_rotation_period, create_new_subnet_with_keys,
+    empty_subnet_update, execute_update_subnet_proposal, get_public_key_with_retries,
+    get_signature_with_logger, verify_signature,
 };
 use ic_management_canister_types::MasterPublicKeyId;
 use ic_nns_constants::GOVERNANCE_CANISTER_ID;
