+++ rs/tests/src/tecdsa/tecdsa_two_signing_subnets_test.rs
@@ -24,10 +24,13 @@ end::catalog[] */
 use crate::orchestrator::utils::subnet_recovery::{
     get_master_public_key, run_chain_key_signature_test,
 };
-use crate::tecdsa::{create_new_subnet_with_keys, make_key, KEY_ID1};
 use anyhow::bail;
 use canister_test::Canister;
 use ic_consensus_system_test_utils::rw_message::install_nns_and_check_progress;
+use ic_consensus_threshold_sig_system_test_utils::{
+    create_new_subnet_with_keys, empty_subnet_update, execute_update_subnet_proposal, make_key,
+    KEY_ID1,
+};
 use ic_management_canister_types::MasterPublicKeyId;
 use ic_nns_constants::GOVERNANCE_CANISTER_ID;
 use ic_registry_subnet_features::{ChainKeyConfig, KeyConfig, DEFAULT_ECDSA_MAX_QUEUE_SIZE};
@@ -42,8 +45,6 @@ use ic_system_test_driver::util::*;
 use ic_types::{Height, SubnetId};
 use registry_canister::mutations::do_update_subnet::UpdateSubnetPayload;
 use slog::{info, Logger};
-
-use super::{empty_subnet_update, execute_update_subnet_proposal};
 
 const NODES_COUNT: usize = 4;
 const DKG_INTERVAL: u64 = 9;
